use skyboJavaFX-Tester - draft-integrating UI wApp
draft, is pulling from the FXML, but not from the other ones. I think right now i have each as its own FXML file for entireFrame, I think that I need to be joining our UI and our skybox scenes and displaying them as one in a pane. will work on it more tomorrow